+++ Changelog.md
@@ -4,6 +4,14 @@ All notable changes to the DBReport project will be documented in this file.
 
 ## [Unreleased]
 
+## [1.23] - 2025-05-10 Pre-release .Net Core in the DotNetCore branch
+### Updated
+- Ngplsql: the open source .NET data provider for PostgreSQL
+- Dynamic configuration using shortcut (e.g. DBReport_Config1.exe.config read instead of DBReport.exe.config) does not work for the moment;
+- The database connection must be explicit at compile time (trying to connect at run time is no longer possible, and DbProviderFactories is no longer supported in App.Config);
+- The package for Oracle.ManagedDataAccess is not available in .Net Core, but in .Net 4 (so the connection constructor is ensured via .Net 4 backward compatibility and not directly in .Net Core);
+- Net 4.8 update to .Net Core (.Net 8)
+
 ## [1.22] - 2025-05-10
 ### Added
 - Option UseUpperCaseIdentifiers for tables, fields and indexes.